requirements.txt
kafka-python
snscrape